@@ -19,8 +19,12 @@
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.10.0",
+        "vue": "^3.2.36",
+        "vue-loader": "^17.0.0",
         "vue-template-compiler": "^2.6.12"
+    },
+    "dependencies": {
+        "autoprefixer": "10.4.7",
+        "vue-router": "^4.0.13"
     }
 }
