@@ -20,7 +20,8 @@
             }
         </style>
     </head>
-    <body class="antialiased">
+    <body class="antialiased" id="app">
+    <search-panel></search-panel>
         <div class="relative flex items-top justify-center min-h-screen bg-gray-100 dark:bg-gray-900 sm:items-center py-4 sm:pt-0">
             @if (Route::has('login'))
                 <div class="hidden fixed top-0 right-0 px-6 py-4 sm:block">
